Change to EmptyPackResources
