+++ src/main/java/net/minecraftforge/common/extensions/IForgePackResources.java
@@ -5,13 +5,6 @@
 
 package net.minecraftforge.common.extensions;
 
-import java.util.Collection;
-import org.jetbrains.annotations.Nullable;
-import net.minecraft.server.packs.PackResources;
-import net.minecraft.server.packs.PackType;
-import net.minecraft.server.packs.resources.ResourceManager;
-import net.minecraftforge.resource.DelegatingPackResources;
-
 public interface IForgePackResources
 {
     /**
@@ -21,14 +14,4 @@ public interface IForgePackResources
     {
         return false;
     }
-
-    /**
-     * Gets a collection of {@code PackResource} instances nested inside this pack.
-     * Used to merge several packs into one entry in the resource pack selection UI without
-     * losing the ability for each pack to return a resource in
-     * {@link ResourceManager#getResourceStack(ResourceLocation)}
-     * @return Collection of nested {@code PackResource}, or null if this pack has no children
-     * @see DelegatingPackResources
-     */
-    @Nullable default Collection<PackResources> getChildren() { return null; }
 }
